rename update_factory_attribute to factories_u
