+++ backend/api/views/__init__.py
@@ -1,4 +1,4 @@
 from .factories_cr import get_nearby_or_create_factories
-from .update_factory_attribute import update_factory_attribute
+from .factories_u import update_factory_attribute
 from .image_c import post_image
 from .factory_image_c import post_factory_image
